Переименовал проекты, исправил название полей в конфиге
