+++ cs/TagCloud.Core/Abstractions/ICenterShiflter.cs
