+++ cs/TagCloud.Core/Abstractions/ISpiral.cs
