+++ cs/TagCloud.Core/Extensions/RectangleExtensions.cs
