+++ cs/TagCloud.Core/Implementations/ArchimedeanSpiral.cs
