+++ cs/TagCloud.Core/Implementations/CenterShifter.cs
