+++ cs/TagCloud.Core/TagCloud.Core.csproj
@@ -4,6 +4,7 @@
         <TargetFramework>net9.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <RootNamespace>TagsCloudLib</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
