+++ cs/TagCloud.Core/Visualizer/TagCloudVisualizer.cs
@@ -7,19 +7,19 @@ public static class TagCloudVisualizer
     public static void DrawRectangles(IEnumerable<Rectangle> rectangles, string outputPath, 
         TagCloudVisualizationConfig config)
     {
-        using var bitmap = new Bitmap(config.Width, config.Height);
+        using var bitmap = new Bitmap(config.CanvasWidth, config.CanvasHeight);
         using var graphics = Graphics.FromImage(bitmap);
 
-        graphics.Clear(config.BackgroundColor);
+        graphics.Clear(config.CanvasBackgroundColor);
 
-        using var pen = new Pen(config.BorderColor, config.BorderThickness);
-        using var brush = new SolidBrush(config.FillColor);
+        using var pen = new Pen(config.ShapeBorderColor, config.ShapeBorderThickness);
+        using var brush = new SolidBrush(config.ShapeFillColor);
 
         foreach (var rect in rectangles)
         {
-            var shifted = Shift(rect, config.Width / 2, config.Height / 2);
+            var shifted = Shift(rect, config.CanvasWidth / 2, config.CanvasHeight / 2);
 
-            if (config.UseFill)
+            if (config.ShapeUseFill)
             {
                 graphics.FillRectangle(brush, shifted);
             }
